+++ src/operation/distinct/test.rs
@@ -60,7 +60,7 @@ async fn build_with_query() {
 #[cfg_attr(feature = "async-std-runtime", async_std::test)]
 async fn build_with_options() {
     let field_name = "field_name".to_string();
-    let max_time = Duration::new(2 as u64, 0);
+    let max_time = Duration::new(2_u64, 0);
     let options: DistinctOptions = DistinctOptions::builder().max_time(max_time).build();
     let ns = Namespace {
         db: "test_db".to_string(),
